requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
channels>=4.0
channels-redis>=4.2
django-cors-headers>=4.3
whitenoise>=6.6
dj-database-url>=2.1
python-dotenv>=1.0
uvicorn>=0.24
gunicorn>=21.2
psycopg2-binary>=2.9
